requirements.txt
pygame==2.5.2
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
pytest==7.4.3
black==23.11.0
mypy==1.7.1 